mv lock to coordinator
