+++ internlm/utils/common.py
@@ -3,13 +3,10 @@
 
 import bisect
 import inspect
-import multiprocessing
 import os
 import random
-import threading
 from contextlib import contextmanager
 from datetime import datetime
-from enum import Enum, unique
 from typing import Union
 
 import numpy as np
@@ -200,61 +197,6 @@ class SingletonMeta(type):
                 len(args) == 0 and len(kwargs) == 0
             ), f"{cls.__name__} is a singleton class and a instance has been created."
         return cls._instances[cls]
-
-
-@unique
-class LockContextType(Enum):
-    """
-    Enum to express the type of the lock
-    """
-
-    THREAD_LOCK = 1
-    PROCESS_LOCK = 2
-
-
-_LOCK_TYPE_MAPPING = {
-    LockContextType.THREAD_LOCK: threading.Lock,
-    LockContextType.PROCESS_LOCK: multiprocessing.Lock,
-}
-
-
-class LockContext(object):
-    """
-    Overview:
-        Generate a LockContext in order to make sure the thread safety.
-    Interfaces:
-        __init__, __enter__, __exit__
-    Example:
-        >>> with LockContext() as lock:
-        >>>     print("Do something here.")
-    """
-
-    def __init__(self, type_: LockContextType = LockContextType.THREAD_LOCK):
-        r"""
-        Overview:
-            init the lock according to given type
-        """
-        self.lock = _LOCK_TYPE_MAPPING[type_]()
-
-    def acquire(self):
-        self.lock.acquire()
-
-    def release(self):
-        self.lock.release()
-
-    def __enter__(self):
-        """
-        Overview:
-            Entering the context and acquire lock
-        """
-        self.lock.acquire()
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        """
-        Overview:
-            Quiting the context and release lock
-        """
-        self.lock.release()
 
 
 def get_megatron_flops(
